==> analysis/01-clean-up-data-for-analysis.Rmd <==
---
title: "01-clean-up-data-for-analysis"
author: "Ellen Coombs"
date: "10/01/2018"
output: html_document
---

```{r setup, include=TRUE}
knitr::opts_chunk$set(echo = FALSE)
```


# R Markdown for stranding project 
## Cleaning stranding data for analysis 

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

## Data cleaning and binding 
This code cleans up the raw data for analysis in the strandings paper 


1. Cleaning and standardising variables in the Natural History Museum (NHM), Cetacean Stranding Investigation Programme (CSIP) datasets for binding 
2. Cleaning and formatting dates in the NHM and CSIP datasets
3. Binding the 2 datasets (NHM + CSIP first, followed by IWDG which was aquired at a later date, see **02-cleaning-and-adding-irish-data.Rmd**)

## Data download 

The NHM data can be directly downloaded from: 
[nhm.ac.uk/data](https://data.nhm.ac.uk/dataset/8c67d782-c01d-4807-8913-f57d0daabf35) 

The CSIP data can be downloaded by going to: 
[ukstrandings.org](http://ukstrandings.org/)

We saved these two datasets as "EDITNHMdata.csv" and "EDITCSIPdata.csv", respectively. This how they are named hereafter.


## Load libraries/packages
```{r}
library(dplyr)
library(tidyr)
library(ggplot2)
library(readr)
```


## Loading the NHM and CSIP data 
```{r}
#Read in raw data for NHM and CSIP
nhm <- read.csv("EDITNHMdata.csv")
#read in raw data 
csip <- read.csv("EDITCSIPdata.csv")
names(csip)
names(nhm)
```

## Initial exploration of data
This initial exploration is valuable for checking which variables we have across the datasets, as well as their class. This is important for binding the datasets later 
```{r}
#What variables have we got? Check the data imported correctly 
names(nhm)
names(csip)
#Check which class variables are (important when binding)
sapply(csip, class)
sapply(nhm, class)
```

## Adding a new source column to the datasets
This code provides a datasource column should it be required for reference
```{r}
#Add sourcecolumn
nhm$Source <- "NHM"
csip$Source <- "CSIP"
```

## Data cleaning 
### Variable names 
The first chunk of code renames some of the column names - making them uniform across the datsets
The second chunk of code cleans up the date format in the NHM dataset and makes date format uniform in both (i.e. NHM and CSIP datasets)

### Date cleaning and uniformity 
Use `mutuate` form the `dplyr` package to change "beg" "mid" "end" and "Summer" "Winter" etc. which appear in some records in the NHM dataset. The dates are mutated as follows: 

Beg and early = 1
Mid = 14
End = 27
wk 1 = 1
wk 2 = 8
wk 3 = 15
wk 4 = 22
Spring = 20 Mar
Summer = 21 Jun
Autumn = 22 Sep
Winter = 21 Dec
Range of dates becomes NA


```{r}
#Cleaning column names to keep and rename where necessary:
#Name Common
#Name Current Sci
#Date (NHM) Date found (CSIP)
#Location
#Grid ref
#County (NHM) Local authority (CSIP); change to county
#Check all names are the same for binding
#Code for renaming if not 
names(csip)
names(nhm)
csip <- dplyr::rename(csip, County = Local.Authority)
csip <- dplyr::rename(csip, Name.Common = Name.common) 
nhm <-dplyr::rename(nhm, Year = year)
csip <- dplyr::rename(csip, Year = year)
csip <- dplyr::rename(csip, S.W.No. = National.Reference)
#Changing dates and names in the selectnhm
#First check the data format 
csip$Date #all dates appear in the same format...
nhm$Date #some dates appear as August 1929 or Summer 1929, code below to clean
#embedding commands
nhm$Date <- gsub("Summer", "1 Jun",
            gsub("Winter", "1 Dec",
            gsub("Spring", "1 Mar",
            gsub("Autumn", "1 Sept",
            gsub("beg", "1",
            gsub("Beg", "1",
            gsub("mid", "14",
            gsub("Mid", "14",
            gsub("end", "27",
            gsub("End", "27",
            gsub("Early", "1",
            gsub("Ely", "1",
            gsub("Early", "1",
            gsub("wk1", "1",
            gsub("wk2", "8",
            gsub("wk3", "15",
            gsub("wk4", "22",
            gsub("Wk4", "22",
                 nhm$Date))))))))))))))))))
#Checking the above has worked
nhm$Date
head(nhm$Date)
tail(nhm$Date)
```


## Selecting specific columns of the dataset for binding 
These are the specific variables of the dataset that we are interested in 

```{r}
#Select specific coloumns for binding 
selectnhm <- select(nhm, S.W.No., Name.Current.Sci, Name.Common, Latitude, Longitude, County, Date, Year)
selectcsip <- select(csip, S.W.No., Name.Current.Sci, Name.Common, Latitude, Longitude, County, Date, Year)
#Checking that the above has worked 
selectnhm$Date
```

## Date format
This code changes the data format in the NHM dataset first and creates and NHM final dataset (nhmfinal). This proves difficult when the dataset spans two centuries, for example the 1900s to the 2000s (see code). The code then formats the CSIP date to the same format (YYYY-MM-DD)
```{r}
#Changing date format in the NHM dataset
library(lubridate) #For dates stuff 
library(dplyr)
#selectnhm <- mutate(selectnhm, Date = dmy(Date))
#Trying this instead - works by changing all 2000s to 1900s
nhmdates <- select(nhm, Date)
selectnhm <- mutate(nhmdates, Date = format(as.Date(Date, "%d-%b-%y"), "19%y-%m-%d"))
selectnhm
#Changing selectnhm to a 'Date' rather than a 'character'
selectnhm <- mutate(selectnhm, Date = as.Date(Date)) 
#Double check that this has worked i.e. the date is now in Date format rather than a chracter 
sapply(selectnhm, class)
#Adding the mutated date column 
#Changing the Latitudes to 'numeric' 
nhmnew <- select(nhm, S.W.No., Name.Current.Sci, Name.Common, Latitude, Longitude, County, Year)
#nhmnew <- mutate(nhmnew, Latitude = as.numeric(Latitude))
nhmfinal <- bind_cols(nhmnew, selectnhm, .id = NULL) #The final nhm dataset 
#Changing CSIP date to YYYYMMMDD
csip$Date
csip <- mutate(csip, Date = dmy(Date))
#Selecting all of the variables 
csipfinal <- select(csip, S.W.No., Name.Current.Sci, Name.Common, Latitude, Longitude, County, Date, Year)
csipfinal
#merging the two datasets
#Checking that both have the same classes - important for binding the columns 
sapply(nhmfinal, class)
sapply(csipfinal, class)
#CSIP Longitude is a factor 
#Change to numeric (as is the same for latitude, and lat/long in the NHM data)
csipfinal$Longitude <- as.numeric(csipfinal$Longitude)
#Now have nhmfinal and csip final 
```

## Merging the two datasets 
This merges the two datasets into a dataset called "cleandatesnames"
```{r}
#Merging the two datasets 
nhmcsip <- bind_rows(nhmfinal, csipfinal)
View(nhmcsip)
#Saving the new dataset 
write.csv(nhmcsip, file = "cleandatesnames.csv") 
```
