==> analysis/04-temporal-analysis-figs-1-3.Rmd <==
---
title: "04-temporal-analysis-figs-1-3"
author: "Ellen Coombs"
date: "10/01/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```


## Plot Figure.1 Different predictors 
This code is looking at some species plots
Splitting data into years and species and then attempting to look at mysticetes seperately 

### Load packages 
```{r}

#Plots of predictors 

library(ggplot2)

#Geom_mean_max plot 
p <- ggplot(data=Geom_mean_max, aes(x=Year, y=Max_K_index, group=1)) +
  geom_line(colour = "gray44") +
  theme_light() + 
  labs(y = "Geomagnetic index (k index)")

#Population plot 
q <- ggplot(data=Population, aes(x=YEAR, y=POPULATION, group=1)) +
  geom_line(colour = "gray44") +
  theme_light() + 
  labs(y = "Human population (millions)", x = "Year")

#SST
r <- ggplot(data=SST_yearly_max, aes(x=year, y=year_max, group=1)) +
  geom_line(colour = "gray44") +
  theme_light() + 
  labs(y = "Maximum sea surface temperature (°C)", x = "Year")+
  theme_light()

#Storms 
s <- ggplot(data=Storm_data, aes(x=Year, y=Storms, group=1)) +
  geom_line(colour = "gray44") +
  theme_light() + 
  labs(y = "Storm count", x = "Year") +
  theme_light()

#NAO_index 
t <- ggplot(data=NAO_index, aes(x=Year, y=NAO_index, group=1)) +
  geom_line(colour = "gray44") +
  theme_light() + 
  labs(y = "North Atlantic Oscillation index", x = "Year") +
  theme_light()

#Fishing catch 
Fishing <- read.csv("Fishing_data_UK.csv")
u <- ggplot(data=Fishing, aes(x=Year, y=Annual.catches..1000.tonnes., group=1)) +
  geom_line(colour = "gray44") +
  theme_light() + 
  labs(y = "Fishing count (1000 tonnes)", x = "Year") +
  theme_light()


#Putting together for the paper
require(gridExtra)
plot1 <- 
plot2 <- qplot(1)
grid.arrange(s, p, r, t, u, q, ncol=3)

```

## Plot Figure.2 UK and Irish species strandings 1913-2015 
This code is looking at some species plots
Splitting data into years and species and then attempting to look at mysticetes seperately 

### Load packages 
```{r}

#This code is looking at some species plots - still messy 
#Splitting data into years and species and then attempting to look at mysticetes seperately 
#Messy variable names - need to clean 

library(dplyr)
library(tidyverse)
library(ggplot2)
library(reshape) 
library(viridis)


UK_and_Irish <- read.csv("UK_and_Irish_strandings.csv")
UK_and_Irish$X <- NULL

#write.csv(cleaneddata, file = "cleaneddata.csv")

#Having a look at how many of each species 
select(UK_and_Irish, Name.Current.Sci)

#Remove unknowns 
Strandings_known_IRL_UK <- UK_and_Irish %>% 
  filter(!(Name.Current.Sci %in% c("Unknown", "Unknown odontocete", "Unknown odontocete ", "Unknown delphinid ",
                                   "Unknown delphinid", "Unknown delphinid ", "Unknown mysticete")))


#'Species' just looking at Name.Current.Sci
Species <- dplyr::count(Strandings_known_IRL_UK, Name.Current.Sci)
#'Speciesyearcount' cleaned data: a count of current scientific name and year 
speciesyearcount <- dplyr::count(Strandings_known_IRL_UK, Name.Current.Sci, Year) %>%
  na.omit()

#This is just species and the year - no counting or sorting 
#speciesyear <- select(cleaneddata, Year, Name.Current.Sci) %>%
  #na.omit()


#Geom_line of all species for every year 
speciesyearcount <- speciesyearcount %>%
  dplyr::rename(Species = Name.Current.Sci) 


p <- ggplot(data = speciesyearcount, aes(x = Year, y = n, colour= Species))+
  theme_bw() +
  labs(x = "Year", y = "Count") +
  geom_line() +
  theme(legend.position="bottom") +
  theme(legend.key.size = unit(1.2, "lines")) +
  scale_fill_viridis(option="viridis") + 
  facet_wrap(~ Species, scales = 'free') +
  theme_minimal() + 
  theme(strip.background = element_rect(colour = "black", fill = "grey40")) +
  theme_bw()

print(p)

#Save as pdf if desired 
ggsave(p, file = "species1.pdf",  height=12, width=20) 
 
 ```

## Fig. 3 - All  strandings plot 
This markdown document requires the output from `03-final-data-clean.Rmd`. This code plots all cetacean strandings from 1913-2015 from the combined dataset (NHM + CSIP + IWDG) and produces Figure.3 from the paper. 

Before being able to plot Figure 3, we need to gather the data for doing so. This dataframe contains Year, Species and total number of strandings for each species for that year. 

## Load packages 
```{r}
library(dplyr)
library(tidyr)
```

## Load data 
Here we are loading the 'UK_and_Irish.sp' dataset (see `03-final-data-clean.Rmd`), this is the final, cleaned dataset with NHM + CSIP + IWDG combined, all rare species and duplicates removed. 

```{r}
#Load data 
UK_and_Irish_sp <- read.csv("UK_and_Irish_sp.csv")
```

## Create a new dataframe 
This code counts the number of each species which strands per year. It adds 0s when there are no strandings for that species 

```{r}
#'Speciesyearcount' cleaned data: a count of current scientific name and year 
#Had to add dplyr:: when I ran this but may be the case that it isn't needed 

speciesyearcount <- dplyr::count(UK_and_Irish_sp, Name.Current.Sci, Year) %>%
  na.omit()

#This is making sure unknowns aren't factors in the data 
speciesyearcount$Name.Current.Sci <- droplevels(speciesyearcount$Name.Current.Sci)

#Adding a 0 to each year without a record 
speciesyearcount <- speciesyearcount %>% 
  complete(Year = seq(min(1913), max(2015), 1L), Name.Current.Sci = unique(speciesyearcount$Name.Current.Sci))

#NAs -> 0 
speciesyearcount[is.na(speciesyearcount)] <- 0

#Changing the name of the dataset 
all_strandings <- speciesyearcount 
#Rename n to "Total_strandings"
all_strandings <- all_strandings %>% 
  rename(Total_strandings = n)
#Reaname Name.Current.Sci to Species 
all_strandings <- all_strandings %>%
  rename(Species = Name.Current.Sci)

#Save dataset 
write.csv(all_strandings, file = "all_strandings.csv")
```

## Plotting all species which strand each year 
### Load packages 
This code is for Figure 3 in the paper 
```{r}
library(plyr)
library(ggplot2)
library(viridis) 
```

### Load data 
```{r}

#load all the strandings data
alls <- read.csv("all_strandings.csv")
```

## Plot data 
This code rearranges the order of the species to put balaenopterids first. 
```{r}
#chance Total_strandings to Total_events to get # events
plotdat <- ddply(alls, .(Year, Species), summarize, total=sum(Total_strandings))
#remove NAs from the data set


#Change the factor levels
#What levels do we have? 
levels(plotdat$Species)

#Specify the factor levels 
plotdat$Species <- factor(c("Balaenoptera acutorostrata","Balaenoptera borealis",
                       "Balaenoptera musculus", "Balaenoptera physalus", 
                       "Delphinus delphis", "Globicephala melas", 
                       "Grampus griseus",  "Hyperoodon ampullatus", 
                       "Kogia breviceps", "Lagenorhynchus acutus", 
                       "Lagenorhynchus albirostris","Megaptera novaeangliae", 
                       "Mesoplodon bidens", "Mesoplodon mirus", "Orcinus orca",
                       "Phocoena phocoena", "Physeter macrocephalus", 
                       "Pseudorca crassidens", "Stenella coeruleoalba",
                       "Tursiops truncatus", "Ziphius cavirostris"))


plotdat$Species <- factor(plotdat$Species, levels = c("Balaenoptera acutorostrata",
                                                      "Balaenoptera borealis",
                                                      "Balaenoptera musculus", 
                                                      "Balaenoptera physalus",
                                                      "Megaptera novaeangliae",
                                                      "Delphinus delphis",
                                                      "Globicephala melas",
                                                      "Grampus griseus",
                                                      "Hyperoodon ampullatus",
                                                      "Kogia breviceps",
                                                      "Lagenorhynchus acutus",
                                                      "Lagenorhynchus albirostris",
                                                      "Mesoplodon bidens",
                                                      "Mesoplodon mirus","Orcinus orca",
                                                      "Phocoena phocoena",
                                                      "Physeter macrocephalus",
                                                      "Pseudorca crassidens",
                                                      "Stenellacoeruleoalba",
                                                      "Tursiops truncatus","Ziphius cavirostris"))

#remove the zeros so they are transparent in plot
plotdat <- plotdat[plotdat$total != 0,]
# reverse sp. order alphabetically
plotdat$Species <- factor(plotdat$Species, levels = rev(levels(plotdat$Species)))

#what should the upper limit be?
#> max(plotdat$total)
#[1] 501

#build the plot
p <- ggplot(plotdat) +
  geom_tile(aes(x=Year, y=Species, fill=total)) +
  scale_x_continuous(expand=c(0,0)) +
  theme_minimal() +
  labs(fill="Individuals", y="") +
  #needs to be on the log scale because of pho^2 :(
  scale_fill_viridis(trans = "log", na.value="white",
                     #501 was from the above commented 'max' code 
                     limits = c(1, 501),
                     breaks = c(1, 10, 20, 50, 100, 250, 500),
                     labels = c(1, 10, 20, 50, 100, 250, 500)) +
  theme(legend.position="bottom", legend.key.width=unit(0.1, "npc"),
        #This makes the text italic 
        axis.text.y=element_text(face="italic")) +
  #Need to bold the balaenopterids 
  theme(axis.text.y=element_text(face=ifelse(levels(plotdat$Species)=="Balaenoptera acutorostrata","bold","italic")))

print(p)
#Save plot 
ggsave(p, filename="Figure3.pdf", width=10, height=6)

```
