==> analysis/05-spatial-analysis-figs-4-5.Rmd <==
---
title: "05-spatial-analysis-figs-4-5"
author: "Ellen Coombs"
date: "10/01/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

## Code for plotting temporal plot (Figure 4) with key events and decadal spatial plots 
This code makes Figure 4. 
The first section of code is needed to make the dataset for Figure 4.

This markdown document requires the output from `03-final-data-clean.Rmd`

## Load packages
```{r}
library(dplyr)
```

## Load and manipulate data for plotting 
```{r}

#Reading in cleaned data with unknowns removed and rare species removed 
#cleaneddata <- read.csv("UK_and_Irish_strandings.csv")
cleaneddata <- read.csv("UK_and_Irish_sp.csv")

latlong <- select(cleaneddata, Year, Name.Current.Sci, Longitude, Latitude)
bphysalus <- filter(latlong, Name.Current.Sci ==  "Balaenoptera physalus")
bactorostrata <- filter(latlong, Name.Current.Sci ==  "Balaenoptera acutorostrata")
bborealis <- filter(latlong, Name.Current.Sci ==  "Balaenoptera borealis")
bmusculus <- filter(latlong, Name.Current.Sci ==  "Balaenoptera musculus")
#unmysticete <- filter(latlong, Name.Current.Sci ==  "Unknown mysticete")
#unbalaenopterid <- filter(latlong, Name.Current.Sci ==  "Unknown balaenoptera")
#mysticete <- filter(latlong, Name.Current.Sci == "Un. mystitcete")
mnovaeangliae <- filter(latlong, Name.Current.Sci == "Megaptera novaeangliae")

#bind all mysticetes
combinedmysticetes <- rbind(bphysalus, bactorostrata, bborealis, bmusculus, mnovaeangliae)

#Stripping out odontocetes 
odontocetes <- latlong[ !(latlong$Name.Current.Sci %in% sortedmysticetes$Name.Current.Sci), ]

#Ordering and combining species by year - to arrange in descending order put "desc" before the year 
sortedodonts <- arrange(odontocetes,(Year))

sortedmysticetes$whatareyou <- "mysticete"
sortedodonts$whatareyou <- "odontocete"


#Beaked whales coded seperately if required 
#beakers <- latlong %>% 
#  filter(Name.Current.Sci %in% c("Hyperoodon ampullatus", "Mesoplodon densirostris", "Mesoplodon europaeus", "Mesoplodon mirus"))
#
#beakers$whatareyou <- "beaker"

allall <- rbind(sortedodonts, sortedmysticetes)
allall$whatareyou <- "all"
allparv <- rbind(sortedodonts, sortedmysticetes, allall)


save(allparv, file="allparv.RData")

```


## Plotting the data 
### Load packages 
```{r}
library(maps)
library(mapdata)
library(ggplot2)
library(viridis)
library(hexbin)
library(ggmap)
library(gridExtra)
library(tidyverse)
```

##Plot Figure.4 with key, mass strandings, and cirles
```{r}

#Required packages 
library(dplyr)
library(ggplot2)
#For adding text underneath the plot 
library(gridExtra)
library(grid)
library(gtable)

#This code plots new Figure 1 (line plot)
#New code on 15/06/2018 with the key underneath 
#Final plot for paper 

#Create text grob 
Text1 = textGrob(paste("
1. 1920s: Sonar use in French and UK waters
2. 1946: NATO military testing in European waters: submarine, sonar, & torpedo testing increase
3. 1950s: Increase in post-war fishing & whaling effort
4. 1960s: Increase in use of polychlorinated biphenyls (PCBs) and other chemical pollutants 
5. 1985/86 season: Moratorium on whaling comes into effect 
6. 1990: The CSIP and IWDG programmes start
7. 2000s: Increase in pile-driving for offshore wind turbines"), 
                       gp = gpar(fontsize = 8, col = 'black'), hjust = 0)

# get the data
linesdat <- ds %>%
  group_by(Year, whatareyou) %>%
  summarise(total=n())


# make the timeseries plot
plines <- ggplot(linesdat) +
  scale_x_continuous(labels=seq(1900, 2025, by=25),
                     breaks=seq(1900, 2025, by=25),
                     expand=c(0, 0)) +
  geom_line(aes(x=Year, y=total, group=whatareyou, colour=whatareyou)) +
  #scale_fill_viridis() +
  #Changed variable colours - these are colour blind friendly 
  scale_colour_manual(values=c("#D55E00", "#0072B2", "#73BFB8")) +
  #position = "jitter" + 
  theme_minimal() +
  theme(legend.position= "right") +
  theme(axis.title.y = element_text(margin = margin(t = 0, r = 15, b = 0, l = 0))) +
  labs(x="", y="Total stranded individuals", colour="Suborder")

#Adding events to the plot 
plines <- plines + annotate("rect", xmin=1914, xmax=1918, ymin=0, ymax=800, alpha=.1, fill="gray22") +
  geom_text(
    #where the label starts (on the x axis) and where it finishes (on the y axis)
    #Colour of the label 
    aes(x = 1916, y = 780, label = "WWI"), size = 3, colour = "gray38") + 
  annotate("rect", xmin=1939, xmax=1945, ymin=0, ymax=800, alpha=.1, fill="gray22") +
  geom_text(
    aes(x = 1942, y = 780, label = "WWII"), size = 3, colour = "gray38") + 
  geom_text(
    aes(x = 1990, y = 230, label = "5"), size = 3, colour = "gray38") +
  geom_text(
    aes(x = 1992, y = 360, label = "6"), size = 3, colour = "gray38") +
  geom_text(
    aes(x = 1953, y = 50, label = "3"), size = 3, colour = "gray38") +
  geom_text(
    aes(x = 1946, y = 50, label = "2"), size = 3, colour = "gray38") +
  geom_text(
    aes(x = 1920, y = 50, label = "1"), size = 3, colour = "gray38") + 
  geom_text(
    aes(x = 2000, y = 500, label = "7"), size = 3, colour = "gray38") + 
  geom_text(
    aes(x = 1960, y = 50, label = "4"), size = 3, colour = "gray38")
  

#Adding circles for mass stranding years 
plines <- plines + geom_point(aes(x=1950, y=275), size=10, shape=1, color="black") + 
  geom_point(aes(x=1927, y=190), size=10, shape=1, color="gray29") +
  geom_point(aes(x=1935, y=121), size=10, shape=1, color="gray29") +
  geom_point(aes(x=1955, y=85), size=10, shape=1, color="gray29") +
  geom_point(aes(x=1983, y=191), size=10, shape=1, color="gray29") +
  geom_point(aes(x=1985, y=106), size=10, shape=1, color="gray29") +
  geom_point(aes(x=2008, y=614), size=10, shape=1, color="gray29") + 
  geom_point(aes(x=2011, y=705), size=10, shape=1, color="gray29") +
  geom_point(aes(x=2012, y=714), size=10, shape=1, color="gray29") +
  geom_point(aes(x=2015, y=667), size=10, shape=1, color="gray29")


#Alter this code to change the positioning of the text grob 
#Add a row below the 2nd from the bottom
gd = ggplotGrob(plines)
gd = gtable_add_rows(gd, unit(1.5, "grobheight", Text1), -3)
#Add 'lab' grob to that row, under the plot panel
gd = gtable_add_grob(gd, Text1, t = -3, l = gd$layout[gd$layout$name == "panel",]$l) 

#Print plot 
grid.newpage()
grid.draw(gd)


#Run the plot - this runs it without the text underneath 
plines

```


## Plot Figure.5 UK and Ireland maps with strandings 

### Load packages 
```{r}
library(ggplot2)
library(viridis)
library(dplyr)
library(maps)
library(hexbin)
library(ggmap)
library(gridExtra)
library(Hmisc) 

#Extract UK map
#Added Ireland
# Note that package purrr masks the function map so detach this, or don't load it, or
# the tidyverse bundle, in the first place. 
uk <- map_data("world", regions = c('UK', 'Ireland', 'Guernsey', 'Jersey', 'Isle of Man'))

#Create base map
gg1 <-
  ggplot() +
  geom_polygon(data = uk, aes(x = long, y = lat, group = group),
               fill = "white", color = "grey70") +
  coord_fixed(1.3)

#Read in the strandings data - this isn't needed here - read in the allparv.RData
#This is the cleaned data with unknowns and rare species removed
#ds <- read.csv("UK_and_Irish_sp.csv")

#This dataset comes from "make_megadata.R"
load("allparv.RData")

ds <- allparv

# capitalize the labels
ds$whatareyou <- Hmisc::capitalize(ds$whatareyou)


#function to build the hexmaps
make_data <- function(ds, years, labs=FALSE){
  # Remove NAs from coordinates
  # And restrict to things in UK waters
  #Filter species here too·
  ds <- ds %>%
    filter(!is.na(Latitude) & !is.na(Longitude)) %>%
    filter(Latitude < 62 & Latitude > 45) %>%
    filter(Longitude < 3 & Latitude > -11)
  
  
  ds <- ds %>%
    filter(Year>=years[1] & Year<years[2])
  
  #Basic plot using viridis colour scheme
  #Note that you can change bins and transparency
  pp <- gg1+
    geom_hex(data = ds, aes(y = Latitude, x= Longitude), bins = 50, alpha = 0.75) +
    theme_minimal() +
    facet_wrap(~whatareyou, ncol=1, strip.position="left") +
    labs(x="", y="") +
    scale_fill_viridis(trans = "log", na.value="white",
                       limits = c(1, 501),
                       breaks = c(1, 10, 20, 50, 100, 250, 500),
                       labels = c(1, 10, 20, 50, 100, 250, 500)) +
    ggtitle(paste0(years[1], "-", years[2])) 

  
  if(labs){
    pp <- pp +
      theme(axis.text=element_blank(),
            strip.text=element_text(),
            plot.title=element_text(hjust=0.5, size=8))
    
  }else{
    pp <- pp +
      theme(legend.position="none",
            strip.text=element_text(colour="white"),
            axis.text=element_blank(),
            plot.title=element_text(hjust=0.5, size=8))
  }
  
  return(pp)
}


  #put the maps together 
# To fix the spacing use plot.margin. Doing it here means you can have different
# margins for each on if needed.
# Top, right, bottom, left is the order
# You may wish to modify. 
  gr <- grid.arrange(make_data(ds, years=c(1913, 1925)) +
                     theme(plot.margin=unit(c(5.5, 5.5, 5.5, -10), "points"),
                           strip.text=element_text(colour="black")), # For y labels
                     make_data(ds, years=c(1926, 1950)) +
                     theme(plot.margin=unit(c(5.5, 5.5, 5.5, -10), "points")),
                     make_data(ds, years=c(1951, 1975)) +
                     theme(plot.margin=unit(c(5.5, 5.5, 5.5, -10), "points")),
                     make_data(ds, years=c(1976, 2000)) +
                     theme(plot.margin=unit(c(5.5, 5.5, 5.5, -10), "points")),
                     make_data(ds, years=c(2001, 2016)) +
                     theme(plot.margin=unit(c(5.5, 5.5, 5.5, -10), "points")),
                     ncol = 5) 
  
# Save to pdf - may want to change size/location/name
  ggsave(gr, file = "megaplot_maps.pdf", height = 5) 
  
```
