==> analysis/12-correlates-data-for-model.Rmd <==
---
title: "11-correlates-data-for-model"
author: "Ellen Coombs"
date: "10/01/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

This code pulls together all of the correlates data e.g. Storms, and NAO, and puts them together along with 'Year', 'Species' and Total strandings' which is called `all_strandings.csv`. This dataset is fed into the Generalised Additive Model (`12-generalised-additive-models.Rmd`)


## Combining correlates data 
This script requires the outputs from: 
`04-temporal-analysis-figure-1.Rmd`
`06-correlates-data-population.Rmd`
`07-correlates-data-storms.Rmd`
`08-correlates-data-geomagnetic.Rmd`
`09-correlates-data-SST.Rmd`
`10-correlates-data-NAO.Rmd`
`11-correlates-data-fishing-catch.Rmd`

## Load packages 
```{r}
library(dplyr)
library(tidyr)
```

## Load correlates data 
```{r}

#Read in species yearly strandings data 
all_strandings <- read.csv("all_strandings.csv")

#Read in UK and Irish population data 
Population <- read.csv("Population_UK.csv")

#Read in storms data 
Storms <- read.csv("Storm_data.csv")

#Read in geomagnetic data 
Final_geom <- read.csv ("Geom_mean_max.csv")
Final_geom$X <- NULL

#Read in SST data 
SST_yearly_max <- read.csv("SST_yearly_max.csv")

#Read in NAO data 
NAO_data <- read.csv("NAO_data.csv")
NAO_data$X <- NULL

#Read in fishing catch data 
Fishing_data <- read.csv("Fisheries_data.csv")

```

## Combining all of the data for final model dataset 
The All_stranding dataset needs to be read back in for all cetacean strandings: this is created in `04-temporal-analysis-figure-1.Rmd`
```{r}

All_model <- bind_cols(Population, Storms, Final_geom, SST_yearly_max, NAO_data, Fishing_data)

Model_data <- All_model
#Save if required 
#write.csv(All_model, file = "Model_data.csv")


#Remove uneeded columns 
Model_data$Year1 <- NULL
Model_data$Year2 <- NULL 
Model_data$year <- NULL 
Model_data$Year3 <- NULL 


#join the two datasets
all_strandings <- full_join(all_strandings, Model_data, by = "Year")
all_strandings$X <- NULL

#Dataset should now have all above correlates + Species + Total_strandings 

```
