==> analysis/13-generalised-additive-models.Rmd <==
---
title: "12-generalised-additive-models""
author: "Ellen Coombs"
date: "10/01/2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

## Running generalised additive models (GAMs) to investigate correlates of strandings
This script requires the output from `12-correlates-data-for-model`

## Load packages 
```{r}
library(dplyr)
library(mgcv)
library(devtools) #for downlading gratia 

#Installing gratia for extra nice plots (if required)
devtools::install_github('gavinsimpson/gratia')
library(gratia)
```

## Load data 
```{r}

all_strandings <- read.csv("all_strandings.csv")

```

## Generalised Additive Models (GAMs)
This code runs the GAM which looks at all strandings (species per year), with an offset of population. The predictors are storms, geomagnetic data (Max_K_index), Maximum SST (Max_SST) and NAO index (NAO_index)

The code also runs the models with harbour porpoises (_phocoena phocoena_) removed, and with harbour porpoises only. 
```{r}

#This is to check how high to make the k value (k-1)
#k (almost always k-1)
unique(all_strandings$Max_SST)
unique(all_strandings$Storms)
unique(all_strandings$Max_K_index)


#GAM for the above with Species as the factor smooth (bs="fs")
#Other variables are thin plate splines (bs = "ts") 
All_strand <- gam(Total_strandings ~ offset(log(Population)) +s(Year, Species, bs="fs") +
                      s(Storms, k=7, bs="ts") +
                      s(Max_K_index, k=4, bs="ts") +
                      s(Max_SST, bs="ts") +
                      s(NAO_index, bs="ts") + 
                      s(Fish_catch, bs = "ts"), 
                    data= all_strandings, 
                    method = "REML",
                    family=nb())

#GAM summary and GAM plots 
summary(All_strand)
par(mfrow = c(2,2))
plot(All_strand)

#AIC if required 
#AIC(All_strand)

#Gam.check
par(mfrow=c(2,2))
gam.check(All_strand)

```


## Tidyng up the outputs for analysis 
```{r}

library(broom) #for getting a tidy data output 

#Tidy gives the neat model output - summarises the model statistical findings 
#e.g. tidy(b_m)
#construct a concise one-row summary of the model. This typically contains values such as R^2, 
#adjusted R^2, and residual standard error that are computed once for the entire model.
#e.g. glance(b_m)

Tidy_All_strand <- list(All_strandc = All_strandc) 

#Saving the tidy and glance datasets 
All_strand_tidy <- plyr::ldply(Tidy_All_strand, tidy, .id = "model")
All_strand_glance <- plyr::ldply(Tidy_All_strand, glance, .id = "model")

#Save to csv if required 
#write.csv(All_strand_tidy, file = "All_strand_tidy.csv")
#write.csv(All_strand_glance, file = "All_strand_glance.csv")
```
